+++ inc/includes/theme-setup.php
@@ -138,6 +138,12 @@ function build_theme_support() {
   add_theme_support( 'wp-block-styles' );
   add_theme_support( 'responsive-embeds' );
 
+  // WooCommerce
+  add_theme_support( 'woocommerce' );
+  add_theme_support( 'wc-product-gallery-zoom' );
+  add_theme_support( 'wc-product-gallery-lightbox' );
+  add_theme_support( 'wc-product-gallery-slider' );
+
   if ( isset( $cp_version ) ) {
     return;
   }